tests: moved all topic creation test to single file
