--- tests/rptest/tests/recreate_topic_metadata_test.py
@@ -1,116 +1,0 @@
-# Copyright 2020 Redpanda Data, Inc.
-#
-# Use of this software is governed by the Business Source License
-# included in the file licenses/BSL.md
-#
-# As of the Change Date specified in that file, in accordance with
-# the Business Source License, use of this software will be governed
-# by the Apache License, Version 2.0
-
-import random
-from ducktape.mark import parametrize
-from rptest.clients.kafka_cat import KafkaCat
-from rptest.clients.rpk import RpkTool
-from rptest.services.admin import Admin
-from rptest.services.cluster import cluster
-from ducktape.utils.util import wait_until
-from rptest.services.rpk_producer import RpkProducer
-
-from rptest.tests.redpanda_test import RedpandaTest
-
-# When quickly recreating topics after deleting them, redpanda's topic
-# dir creation can trip up over the topic dir deletion.  This is not
-# harmful because creation is retried, but it does generate a log error.
-# See https://github.com/redpanda-data/redpanda/issues/5768
-RECREATE_LOG_ALLOW_LIST = ["mkdir failed: No such file or directory"]
-
-
-class RecreateTopicMetadataTest(RedpandaTest):
-    def __init__(self, test_context):
-
-        super(RecreateTopicMetadataTest, self).__init__(
-            test_context=test_context,
-            num_brokers=5,
-            extra_rp_conf={
-                # Test does explicit leadership movements
-                # that the balancer would interfere with.
-                'enable_leader_balancer': False
-            })
-
-    @cluster(num_nodes=6, log_allow_list=RECREATE_LOG_ALLOW_LIST)
-    @parametrize(replication_factor=3)
-    @parametrize(replication_factor=5)
-    def test_recreated_topic_metadata_are_valid(self, replication_factor):
-        """
-        Test recreated topic metadata are valid across all the nodes
-        """
-
-        topic = 'tp-test'
-        partition_count = 5
-        rpk = RpkTool(self.redpanda)
-        kcat = KafkaCat(self.redpanda)
-        admin = Admin(self.redpanda)
-        # create topic with replication factor of 3
-        rpk.create_topic(topic='tp-test',
-                         partitions=partition_count,
-                         replicas=replication_factor)
-
-        # produce some data to the topic
-
-        def wait_for_leader(partition, expected_leader):
-            leader, _ = kcat.get_partition_leader(topic, partition)
-            return leader == expected_leader
-
-        def transfer_all_leaders():
-            partitions = rpk.describe_topic(topic)
-            for p in partitions:
-                replicas = set(p.replicas)
-                replicas.remove(p.leader)
-                target = random.choice(list(replicas))
-                admin.partition_transfer_leadership("kafka", topic, p.id,
-                                                    target)
-                wait_until(lambda: wait_for_leader(p.id, target),
-                           timeout_sec=30,
-                           backoff_sec=1)
-            msg_cnt = 100
-            producer = RpkProducer(self.test_context,
-                                   self.redpanda,
-                                   topic,
-                                   16384,
-                                   msg_cnt,
-                                   acks=-1)
-
-            producer.start()
-            producer.wait()
-            producer.free()
-
-        # transfer leadership to grow the term
-        for i in range(0, 10):
-            transfer_all_leaders()
-
-        # recreate the topic
-        rpk.delete_topic(topic)
-        rpk.create_topic(topic='tp-test',
-                         partitions=partition_count,
-                         replicas=replication_factor)
-
-        def metadata_consistent():
-            # validate leadership information on each node
-            for p in range(0, partition_count):
-                leaders = set()
-                for n in self.redpanda.nodes:
-                    admin_partition = admin.get_partitions(topic=topic,
-                                                           partition=p,
-                                                           namespace="kafka",
-                                                           node=n)
-                    self.logger.info(
-                        f"node: {n.account.hostname} partition: {admin_partition}"
-                    )
-                    leaders.add(admin_partition['leader_id'])
-
-                self.logger.info(f"{topic}/{p} leaders: {leaders}")
-                if len(leaders) != 1:
-                    return False
-            return True
-
-        wait_until(metadata_consistent, 45, backoff_sec=2)
