+++ tests/rptest/tests/topic_creation_test.py
@@ -1,0 +1,309 @@
+# Copyright 2022 Redpanda Data, Inc.
+#
+# Use of this software is governed by the Business Source License
+# included in the file licenses/BSL.md
+#
+# As of the Change Date specified in that file, in accordance with
+# the Business Source License, use of this software will be governed
+# by the Apache License, Version 2.0
+
+from enum import Enum
+import random
+import string
+from time import sleep
+from rptest.clients.default import DefaultClient
+from rptest.services.admin import Admin
+from rptest.services.cluster import cluster
+from rptest.clients.types import TopicSpec
+from rptest.clients.rpk import RpkTool
+from rptest.clients.kafka_cat import KafkaCat
+from rptest.services.producer_swarm import ProducerSwarm
+from rptest.services.redpanda import ResourceSettings, SISettings
+from rptest.services.rpk_producer import RpkProducer
+from rptest.clients.kafka_cli_tools import KafkaCliTools
+
+from ducktape.utils.util import wait_until
+from ducktape.mark import matrix, parametrize
+
+from rptest.tests.redpanda_test import RedpandaTest
+
+
+class Workload():
+    ACKS_1 = 'ACKS_1'
+    ACKS_ALL = 'ACKS_ALL'
+    IDEMPOTENT = 'IDEMPOTENT'
+
+
+class TopicRecreateTest(RedpandaTest):
+    def __init__(self, test_context):
+        super(TopicRecreateTest,
+              self).__init__(test_context=test_context,
+                             num_brokers=5,
+                             resource_settings=ResourceSettings(num_cpus=1),
+                             extra_rp_conf={
+                                 "auto_create_topics_enabled": False,
+                                 "max_compacted_log_segment_size":
+                                 5 * (2 << 20)
+                             })
+
+    @cluster(num_nodes=6)
+    @matrix(
+        workload=[Workload.ACKS_1, Workload.ACKS_ALL, Workload.IDEMPOTENT],
+        cleanup_policy=[TopicSpec.CLEANUP_COMPACT, TopicSpec.CLEANUP_DELETE])
+    def test_topic_recreation_while_producing(self, workload, cleanup_policy):
+        '''
+        Test that we are able to recreate topic multiple times
+        '''
+        self._client = DefaultClient(self.redpanda)
+
+        # scaling parameters
+        partition_count = 30
+        producer_count = 10
+
+        spec = TopicSpec(partition_count=partition_count, replication_factor=3)
+        spec.cleanup_policy = cleanup_policy
+
+        self.client().create_topic(spec)
+
+        producer_properties = {}
+        if workload == Workload.ACKS_1:
+            producer_properties['acks'] = 1
+        elif workload == Workload.ACKS_ALL:
+            producer_properties['acks'] = -1
+        elif workload == Workload.IDEMPOTENT:
+            producer_properties['acks'] = -1
+            producer_properties['enable.idempotence'] = True
+        else:
+            assert False
+
+        swarm = ProducerSwarm(self.test_context,
+                              self.redpanda,
+                              spec.name,
+                              producer_count,
+                              10000000000,
+                              log_level="ERROR",
+                              properties=producer_properties)
+        swarm.start()
+
+        rpk = RpkTool(self.redpanda)
+
+        def topic_is_healthy():
+            partitions = rpk.describe_topic(spec.name)
+            offsets_present = [p.high_watermark > 0 for p in partitions]
+            return len(offsets_present) == partition_count and all(
+                offsets_present)
+
+        for i in range(1, 20):
+            rf = 3 if i % 2 == 0 else 1
+            self.client().delete_topic(spec.name)
+            spec.replication_factor = rf
+            self.client().create_topic(spec)
+            wait_until(topic_is_healthy, 30, 2)
+            sleep(5)
+
+        swarm.stop_all()
+        swarm.wait()
+
+
+class TopicAutocreateTest(RedpandaTest):
+    """
+    Verify that autocreation works, and that the settings of an autocreated
+    topic match those for a topic created by hand with rpk.
+    """
+    def __init__(self, test_context):
+        super(TopicAutocreateTest, self).__init__(
+            test_context=test_context,
+            num_brokers=1,
+            extra_rp_conf={'auto_create_topics_enabled': False})
+
+        self.kafka_tools = KafkaCliTools(self.redpanda)
+        self.rpk = RpkTool(self.redpanda)
+
+    @cluster(num_nodes=1)
+    def topic_autocreate_test(self):
+        auto_topic = 'autocreated'
+        manual_topic = "manuallycreated"
+
+        # With autocreation disabled, producing to a nonexistent topic should not work.
+        try:
+            # Use rpk rather than kafka CLI because rpk errors out promptly
+            self.rpk.produce(auto_topic, "foo", "bar")
+        except Exception:
+            # The write failed, and shouldn't have created a topic
+            assert auto_topic not in self.kafka_tools.list_topics()
+        else:
+            assert False, "Producing to a nonexistent topic should fail"
+
+        # Enable autocreation
+        self.redpanda.set_cluster_config({'auto_create_topics_enabled': True})
+
+        # Auto create topic
+        assert auto_topic not in self.kafka_tools.list_topics()
+        self.kafka_tools.produce(auto_topic, 1, 4096)
+        assert auto_topic in self.kafka_tools.list_topics()
+        auto_topic_spec = self.kafka_tools.describe_topic(auto_topic)
+        assert auto_topic_spec.retention_ms is None
+        assert auto_topic_spec.retention_bytes is None
+
+        # Create topic by hand, compare its properties to the autocreated one
+        self.rpk.create_topic(manual_topic)
+        manual_topic_spec = self.kafka_tools.describe_topic(auto_topic)
+        assert manual_topic_spec.retention_ms == auto_topic_spec.retention_ms
+        assert manual_topic_spec.retention_bytes == auto_topic_spec.retention_bytes
+
+        # Clear name and compare the rest of the attributes
+        manual_topic_spec.name = auto_topic_spec.name = None
+        assert manual_topic_spec == auto_topic_spec
+
+
+class CreateTopicsTest(RedpandaTest):
+
+    #TODO: add shadow indexing properties:
+    #
+    # 'redpanda.remote.write': lambda: random.choice(['true', 'false']),
+    # 'redpanda.remote.read':    lambda: random.choice(['true', 'false'])
+    _topic_properties = {
+        'compression.type':
+        lambda: random.choice(["producer", "zstd"]),
+        'cleanup.policy':
+        lambda: random.choice(["compact", "delete", "compact,delete"]),
+        'message.timestamp.type':
+        lambda: random.choice(["LogAppendTime", "CreateTime"]),
+        'segment.bytes':
+        lambda: random.randint(1024 * 1024, 1024 * 1024 * 1024),
+        'retention.bytes':
+        lambda: random.randint(1024 * 1024, 1024 * 1024 * 1024),
+        'retention.ms':
+        lambda: random.randint(-1, 10000000),
+        'max.message.bytes':
+        lambda: random.randint(1024 * 1024, 10 * 1024 * 1024),
+    }
+
+    def __init__(self, test_context):
+        si_settings = SISettings(
+            cloud_storage_reconciliation_interval_ms=50,
+            cloud_storage_max_connections=5,
+            cloud_storage_segment_max_upload_interval_sec=10,
+            log_segment_size=100 * 1024 * 1024)
+
+        super(CreateTopicsTest, self).__init__(test_context=test_context,
+                                               num_brokers=3,
+                                               si_settings=si_settings)
+
+    def _topic_name(self):
+        return "test-topic-" + "".join(
+            random.choice(string.ascii_lowercase) for _ in range(16))
+
+    @cluster(num_nodes=3)
+    def test_create_topic_with_single_configuration_property(self):
+        rpk = RpkTool(self.redpanda)
+
+        for p, generator in CreateTopicsTest._topic_properties.items():
+            name = self._topic_name()
+            partitions = random.randint(1, 10)
+            property_value = generator()
+            rpk.create_topic(topic=name,
+                             partitions=partitions,
+                             replicas=3,
+                             config={p: property_value})
+
+            cfgs = rpk.describe_topic_configs(topic=name)
+            assert str(cfgs[p][0]) == str(property_value)
+
+
+# When quickly recreating topics after deleting them, redpanda's topic
+# dir creation can trip up over the topic dir deletion.  This is not
+# harmful because creation is retried, but it does generate a log error.
+# See https://github.com/redpanda-data/redpanda/issues/5768
+RECREATE_LOG_ALLOW_LIST = ["mkdir failed: No such file or directory"]
+
+
+class RecreateTopicMetadataTest(RedpandaTest):
+    def __init__(self, test_context):
+
+        super(RecreateTopicMetadataTest, self).__init__(
+            test_context=test_context,
+            num_brokers=5,
+            extra_rp_conf={
+                # Test does explicit leadership movements
+                # that the balancer would interfere with.
+                'enable_leader_balancer': False
+            })
+
+    @cluster(num_nodes=6, log_allow_list=RECREATE_LOG_ALLOW_LIST)
+    @parametrize(replication_factor=3)
+    @parametrize(replication_factor=5)
+    def test_recreated_topic_metadata_are_valid(self, replication_factor):
+        """
+        Test recreated topic metadata are valid across all the nodes
+        """
+
+        topic = 'tp-test'
+        partition_count = 5
+        rpk = RpkTool(self.redpanda)
+        kcat = KafkaCat(self.redpanda)
+        admin = Admin(self.redpanda)
+        # create topic with replication factor of 3
+        rpk.create_topic(topic='tp-test',
+                         partitions=partition_count,
+                         replicas=replication_factor)
+
+        # produce some data to the topic
+
+        def wait_for_leader(partition, expected_leader):
+            leader, _ = kcat.get_partition_leader(topic, partition)
+            return leader == expected_leader
+
+        def transfer_all_leaders():
+            partitions = rpk.describe_topic(topic)
+            for p in partitions:
+                replicas = set(p.replicas)
+                replicas.remove(p.leader)
+                target = random.choice(list(replicas))
+                admin.partition_transfer_leadership("kafka", topic, p.id,
+                                                    target)
+                wait_until(lambda: wait_for_leader(p.id, target),
+                           timeout_sec=30,
+                           backoff_sec=1)
+            msg_cnt = 100
+            producer = RpkProducer(self.test_context,
+                                   self.redpanda,
+                                   topic,
+                                   16384,
+                                   msg_cnt,
+                                   acks=-1)
+
+            producer.start()
+            producer.wait()
+            producer.free()
+
+        # transfer leadership to grow the term
+        for i in range(0, 10):
+            transfer_all_leaders()
+
+        # recreate the topic
+        rpk.delete_topic(topic)
+        rpk.create_topic(topic='tp-test',
+                         partitions=partition_count,
+                         replicas=replication_factor)
+
+        def metadata_consistent():
+            # validate leadership information on each node
+            for p in range(0, partition_count):
+                leaders = set()
+                for n in self.redpanda.nodes:
+                    admin_partition = admin.get_partitions(topic=topic,
+                                                           partition=p,
+                                                           namespace="kafka",
+                                                           node=n)
+                    self.logger.info(
+                        f"node: {n.account.hostname} partition: {admin_partition}"
+                    )
+                    leaders.add(admin_partition['leader_id'])
+
+                self.logger.info(f"{topic}/{p} leaders: {leaders}")
+                if len(leaders) != 1:
+                    return False
+            return True
+
+        wait_until(metadata_consistent, 45, backoff_sec=2)
